tidy pass

+++ include/mods/mod/Note.hpp
@@ -21,7 +21,7 @@ namespace mods
              
              auto getInstrument() const -> size_t;
              auto getPeriod() const -> u16;
-             auto getEffect() const -> u32;
+             auto getEffect() const -> u8;
              auto getEffectArgument() const -> u32;
              
            private:
